chore(tests): colocate run board coverage - move layout behavior tests beside the run board page - move tooltip and card presentation checks - keep application tests focused on app-level workflows
