+++ src/App.test.tsx
@@ -190,105 +190,6 @@ describe("App", () => {
 		await waitFor(() => expect(window.location.pathname).toBe("/"));
 	});
 
-	// Verifies that the run board can switch between vertical columns and horizontal rows.
-	it("switches the run board between vertical and horizontal layouts", async () => {
-		const user = userEvent.setup();
-		render(<App />);
-
-		await user.click(
-			within(screen.getByRole("complementary")).getByRole("button", {
-				name: /运行看板|Run board/,
-			}),
-		);
-
-		const layoutGroup = await screen.findByRole("group", {
-			name: /切换看板布局|Switch run board layout/,
-		});
-		const board = screen.getByTestId("run-board");
-		const verticalButton = within(layoutGroup).getByRole("button", {
-			name: /竖面板|Vertical panel/,
-		});
-		const horizontalButton = within(layoutGroup).getByRole("button", {
-			name: /水平面板|Horizontal panel/,
-		});
-
-		expect(verticalButton).toHaveAttribute("aria-pressed", "true");
-		expect(horizontalButton).toHaveAttribute("aria-pressed", "false");
-		expect(verticalButton).not.toHaveTextContent("竖面板");
-		expect(horizontalButton).not.toHaveTextContent("水平面板");
-		expect(board).toHaveAttribute("data-layout", "vertical");
-
-		await user.click(horizontalButton);
-
-		expect(verticalButton).toHaveAttribute("aria-pressed", "false");
-		expect(horizontalButton).toHaveAttribute("aria-pressed", "true");
-		expect(board).toHaveAttribute("data-layout", "horizontal");
-	});
-
-	// Verifies that both icon-only board layout controls expose their meaning.
-	it("describes both run board layout controls on hover", async () => {
-		const user = userEvent.setup();
-		render(<App />);
-
-		await user.click(
-			within(screen.getByRole("complementary")).getByRole("button", {
-				name: /运行看板|Run board/,
-			}),
-		);
-
-		const layoutGroup = await screen.findByRole("group", {
-			name: /切换看板布局|Switch run board layout/,
-		});
-		const verticalButton = within(layoutGroup).getByRole("button", {
-			name: /竖面板|Vertical panel/,
-		});
-		const horizontalButton = within(layoutGroup).getByRole("button", {
-			name: /水平面板|Horizontal panel/,
-		});
-
-		await user.hover(verticalButton);
-		const verticalTooltip = await screen.findByRole("tooltip");
-		expect(verticalTooltip).toHaveTextContent(/竖面板|Vertical panel/);
-		expect(verticalTooltip).toHaveClass("whitespace-nowrap");
-		expect(verticalTooltip).toHaveClass("max-w-none");
-
-		await user.unhover(verticalButton);
-		await user.hover(horizontalButton);
-		const horizontalTooltip = await screen.findByRole("tooltip");
-		expect(horizontalTooltip).toHaveTextContent(/水平面板|Horizontal panel/);
-		expect(horizontalTooltip).toHaveClass("whitespace-nowrap");
-		expect(horizontalTooltip).toHaveClass("max-w-none");
-	});
-
-	// Verifies that run cards keep a stable footprint and clamp overflowing copy.
-	it("keeps run cards at fixed dimensions with clamped text", async () => {
-		const user = userEvent.setup();
-		render(<App />);
-
-		await user.click(
-			within(screen.getByRole("complementary")).getByRole("button", {
-				name: /运行看板|Run board/,
-			}),
-		);
-
-		const card = (await screen.findAllByRole("article"))[0];
-
-		expect(card).toHaveClass(
-			"h-48",
-			"w-[18rem]",
-			"max-w-full",
-			"overflow-hidden",
-		);
-		expect(card.querySelector("h3")).toHaveClass(
-			"line-clamp-2",
-			"overflow-hidden",
-		);
-		expect(card.querySelector("p")).toHaveClass(
-			"line-clamp-2",
-			"overflow-hidden",
-		);
-	});
-
 	// Verifies that the running process state refreshes without reloading the UI.
 	it("refreshes the selected agent process state", async () => {
 		let processProbeCount = 0;
